Use a fallback when dealing with unknown types
When we encounter an unknown/unregistered type, simply supply an empty type to the
marshaler, so that it will do the right thing by outputting unknown fields as JSON
map fields.

Before we would get, when the response contained an unregistered type:
Error: marshal unknown message type "<message name>"

Now we get instead the a proper JSON object.